Moved loadable to the markdown editor
