+++ client/coral-embed-stream/src/tabs/configure/components/QuestionBoxBuilder.js
@@ -1,11 +1,56 @@
-import { Spinner } from 'coral-ui';
-import Loadable from 'react-loadable';
+import React from 'react';
+import QuestionBox from '../../../components/QuestionBox';
+import DefaultQuestionBoxIcon from '../../../components/DefaultQuestionBoxIcon';
+import cn from 'classnames';
+import styles from './QuestionBoxBuilder.css';
+import { Icon } from 'coral-ui';
+import MarkdownEditor from 'coral-framework/components/MarkdownEditor';
 
-const QuestionBoxBuilder = Loadable({
-  loader: () =>
-    import(/* webpackChunkName: "questionBoxBuilder" */
-    './loadable/QuestionBoxBuilder'),
-  loading: Spinner,
-});
+const DefaultIcon = <DefaultQuestionBoxIcon className={styles.defaultIcon} />;
+const icons = [{ default: DefaultIcon }, 'forum', 'build', 'format_quote'];
+
+class QuestionBoxBuilder extends React.Component {
+  render() {
+    const {
+      questionBoxIcon,
+      questionBoxContent,
+      onContentChange,
+      onIconChange,
+    } = this.props;
+
+    return (
+      <div className={styles.root}>
+        <h4>Include an Icon</h4>
+
+        <ul className={styles.iconList}>
+          {icons.map(item => {
+            const name = typeof item === 'object' ? Object.keys(item)[0] : item;
+            const icon = typeof item === 'object' ? item[name] : item;
+            return (
+              <li className={styles.item} key={name}>
+                <button
+                  className={cn(styles.button, {
+                    [styles.buttonActive]: questionBoxIcon === name,
+                  })}
+                  onClick={() => onIconChange(name)}
+                >
+                  {typeof icon === 'string' ? <Icon name={icon} /> : icon}
+                </button>
+              </li>
+            );
+          })}
+        </ul>
+
+        <QuestionBox
+          className={styles.questionBox}
+          icon={questionBoxIcon}
+          content={questionBoxContent}
+        />
+
+        <MarkdownEditor value={questionBoxContent} onChange={onContentChange} />
+      </div>
+    );
+  }
+}
 
 export default QuestionBoxBuilder;
